refactor: modify products repository and service to use model
