+++ src/api/products/products.service.ts
@@ -1,16 +1,16 @@
-import { getProducts, getProductById } from "./products.repository";
-import { BadRequestError } from "../../utils/errors";
+import { getAll, getById } from "./products.repository";
+import { BadRequestError } from "../utils/errors";
 
-import type { Product, ProductId } from "./product/product.types";
+import type { ProductId } from "./product/product.types";
 
-export const getProductsService = async (): Promise<Product[]> => {
-  return await getProducts();
+export const getProductsService = async () => {
+  return await getAll()
 };
 
 export const getProductByIdService = async (
   productId: ProductId
-): Promise<Product> => {
-  const product = await getProductById(productId);
+) => {
+  const product = await getById(productId)
   if (!product) {
     throw new BadRequestError(`No product with such id`);
   }
